L641: rb_float_new(*((float*)ptr)) -> rb_float_new(*((double*)ptr)).
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@2418 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ext/dl/ptr.c
@@ -638,7 +638,7 @@ cary2ary(void *ptr, char t, int len)
       ptr = (char *)ptr + sizeof(float);
       break;
     case 'D':
-      elem = rb_float_new(*((float*)ptr));
+      elem = rb_float_new(*((double*)ptr));
       ptr = (char *)ptr + sizeof(double);
       break;
     case 'C':
